tracing, metrics

+++ internal/log/segment.go
@@ -120,7 +120,7 @@ func (s *segment) Close() error {
 //
 // eg. nearestMultiple(9, 4) == 8
 func nearestMultiple(j, k uint64) uint64 {
-	if j >= 0 {
+	if j >= uint64(0) {
 		return (j / k) * k
 	}
 	return ((j - k + 1) / k) * k
